move root_inc_file to jules script
