+++ run_configs/jules/build_jules.py
@@ -6,8 +6,12 @@
 ##############################################################################
 import logging
 import os
+import shutil
+import warnings
+from pathlib import Path
 
 from fab.build_config import BuildConfig
+from fab.steps import step
 from fab.steps.analyse import analyse
 from fab.steps.archive_objects import ArchiveObjects
 from fab.steps.cleanup_prebuilds import CleanupPrebuilds
@@ -17,8 +21,7 @@ from fab.steps.grab.fcm import fcm_export
 from fab.steps.grab.prebuild import GrabPreBuild
 from fab.steps.link import LinkExe
 from fab.steps.preprocess import fortran_preprocessor
-from fab.steps.root_inc_files import root_inc_files
-from fab.util import common_arg_parser
+from fab.util import common_arg_parser, suffix_filter
 
 logger = logging.getLogger('fab')
 
@@ -61,6 +64,48 @@ def jules_config(revision=None, compiler=None, two_stage=False):
     return config
 
 
+@step
+def root_inc_files(config):
+
+    """
+    Copy inc files into the workspace output root.
+
+    Checks for name clash. This step does not create any artefacts.
+    It's up to the user to configure other tools to find these files.
+
+    :param artefact_store:
+        Artefacts created by previous Steps.
+        This is where we find the artefacts to process.
+    :param config:
+        The :class:`fab.build_config.BuildConfig` object where we can read settings
+        such as the project workspace folder or the multiprocessing flag.
+
+    """
+
+    # todo: make the build output path a getter calculated in the config?
+    build_output: Path = config.build_output
+    build_output.mkdir(parents=True, exist_ok=True)
+
+    warnings.warn("RootIncFiles is deprecated as .inc files are due to be removed.", DeprecationWarning)
+
+    # inc files all go in the root - they're going to be removed altogether, soon
+    inc_copied = set()
+    for fpath in suffix_filter(config._artefact_store["all_source"], [".inc"]):
+
+        # don't copy from the output root to the output root!
+        # this is currently unlikely to happen but did in the past, and caused problems.
+        if fpath.parent == build_output:
+            continue
+
+        # check for name clash
+        if fpath.name in inc_copied:
+            raise FileExistsError(f"name clash for inc file: {fpath}")
+
+        logger.debug(f"copying inc file {fpath}")
+        shutil.copy(fpath, build_output)
+        inc_copied.add(fpath.name)
+
+
 if __name__ == '__main__':
     arg_parser = common_arg_parser()
     arg_parser.add_argument('--revision', default=os.getenv('JULES_REVISION', 'vn6.3'))
@@ -70,9 +115,11 @@ if __name__ == '__main__':
     # this contains some of the stuff that was in run(), which needs to come before the steps.
     config._run_prep()
 
+    # grab the source
     fcm_export(config, src='fcm:jules.xm_tr/src', revision=args.revision, dst_label='src')
     fcm_export(config, src='fcm:jules.xm_tr/utils', revision=args.revision, dst_label='utils')
 
+    # find the source files
     find_source_files(config, path_filters=[
         Exclude('src/control/um/'),
         Exclude('src/initialisation/um/'),
@@ -81,6 +128,7 @@ if __name__ == '__main__':
         Exclude('src/params/shared/cable_maths_constants_mod.F90'),
     ])
 
+    # move inc files to the root for easy tool use
     root_inc_files(config)
 
     fortran_preprocessor(
